Remove MU layout from addon dummy app
