+++ packages/@ember/octane-addon-blueprint/files/tests/dummy/app/index.html
