+++ packages/@ember/octane-addon-blueprint/files/tests/dummy/app/resolver.js
@@ -1,0 +1,3 @@
+import Resolver from 'ember-resolver';
+
+export default Resolver;
