+++ packages/@ember/octane-addon-blueprint/files/tests/dummy/app/router.js
@@ -1,5 +1,5 @@
 import EmberRouter from "@ember/routing/router";
-import config from "../config/environment";
+import config from "./config/environment";
 
 const Router = EmberRouter.extend({
   location: config.locationType,
