+++ packages/@ember/octane-addon-blueprint/files/tests/dummy/app/routes/application/template.hbs
@@ -1,0 +1,3 @@
+<h2 id="title">Welcome to Ember</h2>
+
+{{outlet}}
