+++ packages/@ember/octane-addon-blueprint/files/tests/dummy/app/templates/application.hbs
