--- packages/@ember/octane-addon-blueprint/files/tests/dummy/src/resolver.js
@@ -1,13 +1,0 @@
-import Resolver from 'ember-resolver/resolvers/fallback';
-import buildResolverConfig from 'ember-resolver/ember-config';
-import config from '../config/environment';
-
-let moduleConfig = buildResolverConfig(config.modulePrefix);
-/*
- * If your application has custom types and collections, modify moduleConfig here
- * to add support for them.
- */
-
-export default Resolver.extend({
-  config: moduleConfig
-});
